refactor entity (int to Long)

+++ src/main/java/pl/grabowski_durka/dto/OwnerDto.java
@@ -15,7 +15,7 @@ public class OwnerDto {
     private Long id;
     private String name;
     private String surname;
-    private int pesel;
+    private Long pesel;
     private String email;
-    private int phone;
+    private Long phone;
 }
